fix modal edition copy
