+++ src/views/home/components/EditionsPublicList.vue
@@ -10,9 +10,10 @@
                 v-for="edition in editions"
                 :key="edition.versionId"
             >
-                <edition-public-card :edition="edition"></edition-public-card>
+                <edition-public-card @edition-copy-click="openCopyEditionModal(edition)" :edition="edition"></edition-public-card>
             </b-card>
         </div>
+        <copy-edition-modal />
     </div>
 </template>       
 
@@ -22,17 +23,25 @@ import { EditionInfo } from '@/models/edition';
 import EditionIcons from '@/components/cues/edition-icons.vue';
 import Waiting from '@/components/misc/Waiting.vue';
 import EditionPublicCard from './EditionPublicCard.vue';
+import CopyEditionModal from './CopyEditionModal.vue';
 
 @Component({
     name: 'editions-public-list',
     components: { 
         Waiting,
         EditionPublicCard,
+        CopyEditionModal
      },
 })
 export default class EditionsPublicList extends Vue {
     @Prop(
     ) public editions!: EditionInfo[];
+
+        private openCopyEditionModal(edition: EditionInfo) {
+            // console.log(edition);
+        this.$state.editions.current = edition;
+        this.$root.$emit('bv::show::modal', 'copy-edition-modal');
+    }
 }
 </script>
